Make a tablesaw-smile module
