+++ aggregate/pom.xml
@@ -78,5 +78,10 @@
       <artifactId>tablesaw-jsplot</artifactId>
       <version>0.31.1-SNAPSHOT</version>
     </dependency>
+    <dependency>
+      <groupId>tech.tablesaw</groupId>
+      <artifactId>tablesaw-smile</artifactId>
+      <version>0.31.1-SNAPSHOT</version>
+    </dependency>
   </dependencies>
 </project>
